+++ library/DynamicFormItemControls/FormGroup.ts
@@ -13,4 +13,19 @@ export interface FormGroupProps extends HTMLAttributes {
    * flex 布局下的水平排列方式：
    */
   justify: 'start'|'end'|'center'|'space-around'|'space-between';
+  /**
+   * 是否为朴素样式
+   * @default false
+   */
+  plain: boolean,
+  /**
+   * 是否默认折叠
+   * @default false
+   */
+  collapsed: boolean,
+  /**
+   * 是否可折叠
+   * @default false
+   */
+  collapsible: boolean,
 }