+++ library/DynamicFormItemControls/FormGroup.vue
@@ -1,16 +1,28 @@
 <template>
-  <div class="dynamic-form-group">
-    <h5 v-if="title">{{ title }}</h5>
-    <Row :justify="(justify as any)" :gutter="gutter">
+  <div :class="[
+    'dynamic-form-group', 
+    {
+      'collapsed': collapsed,
+      'collapsible': collapsible,
+      'plain': plain,
+    }
+  ]">
+    <h5 v-if="title" @click="collapsible ? collapsed = !collapsed : null">
+      <span>{{ title }}</span>
+      <IconDown v-if="collapsible" class="collapsible-icon" />
+    </h5>
+    <Row v-if="!collapsed" :justify="(justify as any)" :gutter="gutter">
       <slot />
     </Row>
   </div>
 </template>
 
 <script lang="ts" setup>
+import { ref } from "vue";
 import Row from "../DynamicFormBasicControls/Layout/Row.vue";
+import IconDown from "../Images/IconDown.vue";
 
-defineProps({
+const props = defineProps({
   /**
    * 标题
    */
@@ -32,7 +44,32 @@ defineProps({
     type: String,
     default: "start",
   },
-})
+  /**
+   * 是否可折叠
+   */
+  collapsible: {
+    type: Boolean,
+    default: false,
+  },
+  /**
+   * 是否为朴素样式
+   */
+  plain: {
+    type: Boolean,
+    default: false,
+  },
+  /**
+   * 是否默认折叠
+   */
+  collapsed: {
+    type: Boolean,
+    default: false,
+  },
+});
+
+const collapsed = ref(props.collapsed);
+
+
 </script>
 
 <style lang="scss">
@@ -41,10 +78,38 @@ defineProps({
   background-color: var(--dynamic-form-background-color);
   border-radius: var(--dynamic-form-border-radius);
 
+  &.collapsed {
+    padding: 0;
+
+    .collapsible-icon {
+      transform: rotate(0deg);
+    }
+  }
+  &.collapsible {
+    h5 {
+      cursor: pointer;
+    }
+  }
+
+  .collapsible-icon {
+    transform: rotate(180deg);
+    transition: transform 0.3s ease-in-out;
+    width: 16px;
+    height: 16px;
+  }
+
   h5 {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
     color: var(--dynamic-form-text-color);
     margin: 0;
     margin-bottom: 12px;
+  }
+
+  &.plain {
+    padding: 0;
+    background-color: transparent;
   }
 }
 </style>